refractor: format code and remove comment and update package

+++ Application/Features/Statistical/Queries/GetAdminProductReport/GetAdminProductReportQueryHandler.cs
@@ -17,13 +17,14 @@ public sealed class GetAdminProductReportQueryHandler(IStatisticalReadRepository
         var bestSeller = tableData.OrderByDescending(p => p.SoldCount30Days).FirstOrDefault();
         var deadStockOptions = tableData.Where(p => p.StockQuantity > 0).OrderBy(p => p.SoldCount30Days).ToList();
         var deadStock = deadStockOptions.FirstOrDefault();
-        
-        decimal deadStockValue = deadStock != null ? (decimal)(deadStock.StockQuantity * deadStock.SellPrice) : 0; // Ước tính bằng SellPrice vì DTO không có CostPrice, hoặc ta có thể coi đây là giá trị vốn nếu có. Ở Frontend hiển thị "Giá trị".
+
+        decimal deadStockValue = deadStock != null ? deadStock.StockQuantity * deadStock.SellPrice : 0;
 
         double avgTurnover = 0;
-        if (tableData.Any(p => p.StockQuantity > 0))
+        if(tableData.Any(p => p.StockQuantity > 0))
         {
-            avgTurnover = tableData.Where(p => p.StockQuantity > 0).Average(p => (double)p.SoldCount30Days / p.StockQuantity);
+            avgTurnover = tableData.Where(p => p.StockQuantity > 0)
+                .Average(p => (double)p.SoldCount30Days / p.StockQuantity);
         }
 
         var highlights = new ProductReportHighlightsResponse
@@ -39,21 +40,25 @@ public sealed class GetAdminProductReportQueryHandler(IStatisticalReadRepository
         var topRevenue = tableData
             .OrderByDescending(p => p.SoldCount30Days * p.SellPrice)
             .Take(5)
-            .Select(p => new TopProductRevenueResponse
-            {
-                ProductName = p.ProductName,
-                UnitsSold = p.SoldCount30Days,
-                Revenue = p.SoldCount30Days * p.SellPrice
-            }).ToList();
+            .Select(
+                p => new TopProductRevenueResponse
+                {
+                    ProductName = p.ProductName,
+                    UnitsSold = p.SoldCount30Days,
+                    Revenue = p.SoldCount30Days * p.SellPrice
+                })
+            .ToList();
 
         var topProfit = tableData
             .OrderByDescending(p => p.SoldCount30Days * p.SellPrice * (decimal)(p.MarginPercentage / 100))
             .Take(5)
-            .Select(p => new TopProductProfitResponse
-            {
-                ProductName = p.ProductName,
-                Profit = p.SoldCount30Days * p.SellPrice * (decimal)(p.MarginPercentage / 100)
-            }).ToList();
+            .Select(
+                p => new TopProductProfitResponse
+                {
+                    ProductName = p.ProductName,
+                    Profit = p.SoldCount30Days * p.SellPrice * (decimal)(p.MarginPercentage / 100)
+                })
+            .ToList();
 
         return new AdminProductReportResponse
         {
